Completed the GPU AI Node Logic
After all this time, finished writing the logic for the GPU AI Nodes. They needed to sum up the weights and inputs, then AFTER all that is calculated, apply a bias, then finally feed the result through an activation function of some sort (The magic logic that makes the AI capable of non-linear calculation and learning) in order to get the final output. Still crashes, but the complicated node processing part is done!